store task output data basestorage,storage
